requirements.txt
streamlit==1.52.2
openai==1.15.0
openai-whisper
torch
ffmpeg-python
